updated boringssl to newer version linux_sgx_2_9.patch now does not use rand_r anymore fixed bazelversion and added bazelignore for build folders trying to fix a bug in preprocess.py that causes a narrowing in c++ code

+++ asylo/platform/system_call/preprocess.py
@@ -175,8 +175,8 @@ class SystemCallTable(object):
         self.annotation_list:
       key = '{{"{}", "{}"}}'.format(syscall, param_name)
       if annotation_name == 'bound':
-        bind_param_index = self.parameter_list[syscall].index(
-            annotation_value) / 2
+        bind_param_index = int(self.parameter_list[syscall].index(
+            annotation_value) / 2)
         bounds.append('{{{}, {}}}'.format(key, bind_param_index))
       if annotation_name == 'count':
         counts.append('{{{}, {}}}'.format(key, annotation_value))
